Actualizacion del proyecto final de moviles

+++ AndroidStudioProjects/RestaurantesAPI/app/src/main/java/com/example/restaurantesapi/models/Plate.kt
@@ -6,7 +6,7 @@ data class Plate (
     val name: String,
     val description: String,
     val price: String,
-    val menu_category_id: Long
+    val menu_category_id: Int
 ){
     var id: Int? = null
 }
